hw_3 rename file
